# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Keldris - Keeper of your data. Self-hosted backup solution with OIDC authentication, Restic backup engine, and cloud storage support.",
        "title": "Keldris API",
        "termsOfService": "https://keldris.io/terms",
        "contact": {
            "name": "Keldris Support",
            "url": "https://github.com/MacJediWizard/keldris"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/schedules": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns all backup schedules for the current organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "List schedules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by agent ID",
                        "name": "agent_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Schedule"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Creates a new backup schedule for an agent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Create schedule",
                "parameters": [
                    {
                        "description": "Schedule details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Schedule"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/schedules/{id}/run": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Triggers an immediate backup for the specified schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Run schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/handlers.RunScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateScheduleRequest": {
            "type": "object",
            "required": [
                "agent_id",
                "cron_expression",
                "name",
                "paths",
                "repository_id"
            ],
            "properties": {
                "agent_id": {
                    "type": "string"
                },
                "backup_window": {
                    "$ref": "#/definitions/models.BackupWindow"
                },
                "bandwidth_limit_kb": {
                    "type": "integer"
                },
                "cron_expression": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "excluded_hours": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "excludes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "paths": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "repository_id": {
                    "type": "string"
                },
                "retention_policy": {
                    "$ref": "#/definitions/models.RetentionPolicy"
                }
            }
        },
...
        "handlers.RunScheduleResponse": {
            "type": "object",
            "properties": {
                "backup_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.BackupWindow": {
            "type": "object",
            "properties": {
                "end": {
                    "description": "HH:MM format (e.g., \"06:00\")",
                    "type": "string"
                },
                "start": {
                    "description": "HH:MM format (e.g., \"02:00\")",
                    "type": "string"
                }
            }
        },
...
        "models.RetentionPolicy": {
            "type": "object",
            "properties": {
                "keep_daily": {
                    "type": "integer"
                },
                "keep_hourly": {
                    "type": "integer"
                },
                "keep_last": {
                    "type": "integer"
                },
                "keep_monthly": {
                    "type": "integer"
                },
                "keep_weekly": {
                    "type": "integer"
                },
                "keep_yearly": {
                    "type": "integer"
                }
            }
        },
        "models.Schedule": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "type": "string"
                },
                "backup_window": {
                    "description": "Allowed backup time window",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.BackupWindow"
                        }
                    ]
                },
                "bandwidth_limit_kb": {
                    "description": "Upload limit in KB/s",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "cron_expression": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "excluded_hours": {
                    "description": "Hours (0-23) when backups should not run",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "excludes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repository_id": {
                    "type": "string"
                },
                "retention_policy": {
                    "$ref": "#/definitions/models.RetentionPolicy"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Agent API key authentication. Use format: Bearer kld_xxx",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie authentication (for web UI)",
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}